refactor: reorganize folder structure (no logic changes) (#11)
