+++ docs/archiecture.md
@@ -1,17 +1,21 @@
-### 파일구조 
+### 파일구조
 ainsight-backend/
+├── .claude/
+│   └── settings.local.json
 ├── docker/
 │   ├── ingest.Dockerfile
 │   └── web.Dockerfile
 ├── docker-compose.yml
 ├── docs/
+│   ├── archiecture.md
 │   ├── git-branch-rules.md
-│   └── git-message-rules.md
+│   ├── git-message-rules.md
+│   └── github-api-docs.md
 ├── env/
 │   ├── ingest.env
 │   ├── mongo.env
 │   ├── web.env
-│   └── web.dev.env
+│   └── woker.env
 ├── README.md
 ├── requirements/
 │   ├── base.txt
@@ -36,9 +40,9 @@ ainsight-backend/
 │   │   │   ├── __init__.py
 │   │   │   ├── __pycache__/
 │   │   │   └── app_containers.py
-│   │   └── db/
+│   │   └── logging/
 │   │       ├── __init__.py
-│   │       └── mongo.py
+│   │       └── logger.py
 │   ├── domains/
 │   │   └── tmp/
 │   │       ├── models.py
@@ -47,15 +51,23 @@ ainsight-backend/
 │   └── ingest/
 │       ├── __init__.py
 │       ├── main.py
+│       ├── mappers/
+│       │   ├── __init__.py
+│       │   └── github_repo_mapper.py
 │       ├── models/
 │       │   └── document.py
 │       └── sources/
 │           ├── __init__.py
 │           ├── github/
 │           │   ├── client.py
-│           │   └── fetcher.py
-│           │   └── filter.py
-│           │   └── db_indexes.js
+│           │   ├── fetcher.py
+│           │   ├── filters.py
+│           │   ├── indexes.py
+│           │   ├── job_generator.py
+│           │   ├── job_indexes.py
+│           │   ├── job_monitor.py
+│           │   ├── job_schema.py
+│           │   └── job_worker.py
 │           └── reddit/
 │               ├── __init__.py
 │               ├── client.py
